requirements.txt
imageio==2.9.0
numpy==1.18.0
pytorch_msssim==0.2.1
tqdm==4.36.1
torchvision==0.9.0
scikit_image==0.15.0
torch==1.8.0
PyYAML==5.4.1
scikit-image
